backend/requirements.render.txt
fastapi==0.128.0
uvicorn==0.35.0
python-multipart
pydantic==2.12.5
opencv-python-headless
mediapipe==0.10.9
protobuf<4.0.0
numpy
pyserial
requests
python-dotenv
onnxruntime==1.22.1
